Split stl2/memory.hpp into two functions to correct some dependent compile-time errors
